Move appMenu to appBar
